+++ Frends.Confluence.Request/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [1.1.0] - 2024-06-28
+
+### Added
+
+- Added methods: CreatePage, UpdatePage, GetPageById, DeletePage, GetPageByTitle, CreateSpace, DeleteSpace, GetSpaceByName.
+
+### Changed
+
+- Changed Request method name to CustomRequest.
+
 ## [1.0.0] - 2024-05-21
 
 ### Added